src/pages/results/utils/index.ts
import { Order } from '../types';

export function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
  console.log({ a, b, orderBy });
  if (b[orderBy] < a[orderBy]) {
    return -1;
  }
  if (b[orderBy] > a[orderBy]) {
    return 1;
  }
  return 0;
}

export function getComparator<Key extends keyof any>(
  order: Order,
  orderBy: Key
): (a: { [key in Key]?: string }, b: { [key in Key]?: string }) => number {
  return order === 'desc'
    ? (a, b) => descendingComparator(a, b, orderBy)
    : (a, b) => -descendingComparator(a, b, orderBy);
}
